Переносит функции рендера в отдельный файл
