+++ src/components/board.js
@@ -1,4 +1,4 @@
-import {createElement} from "../utils/common";
+import {createElement} from "../utils/render";
 
 const createBoard = () => `<ul class="trip-days"></ul>`;
