+++ src/components/event-edit.js
@@ -1,7 +1,7 @@
 import {createOffers} from "./offers";
 import {createDestination} from "./destination";
 import {createHeader} from "./header-event";
-import {createElement} from "../utils/common";
+import {createElement} from "../utils/render";
 
 const createEventEdit = (event) => {
   const {typeEvent, timeStart, options, destinations, pictures} = event;
