+++ src/components/filter.js
@@ -1,4 +1,4 @@
-import {createElement} from "../utils/common";
+import {createElement} from "../utils/render";
 
 const createFilter = () => {
   return (
